requirements.txt
flask
openai
python-dotenv
pyaudio
webrtcvad